Reorganize API routes into logical domain-based structure
- Restructure API into 8 domain-based routers: agents, messaging, media, memory, audio, runtime, tee, system
- Create shared utilities for response handling, validation, and file operations
- Remove 2,800+ lines of duplicate and legacy code (agent.ts, messages.ts, env.ts, tee.ts, world.ts)
- Maintain backward compatibility through proper domain organization
- Clean up main router integration and remove duplicate route aliases
- Fix TypeScript build errors and import path issues

New structure:
- /agents: Agent CRUD operations and management
- /messaging: Message handling, channels, and communication
- /media: File uploads and media management
- /memory: Memory and knowledge operations
- /audio: Speech, TTS, and audio processing
- /runtime: Server runtime operations (health, logs, stop)
- /tee: TEE security operations
- /system: Environment and system configuration
