+++ packages/cli/src/server/api/agents/panels.ts
@@ -1,0 +1,46 @@
+import type { IAgentRuntime, UUID } from '@elizaos/core';
+import { validateUuid, logger } from '@elizaos/core';
+import express from 'express';
+import type { AgentServer } from '../../index';
+import { sendError, sendSuccess } from '../shared/response-utils';
+
+/**
+ * Agent panels and plugin routes management
+ */
+export function createAgentPanelsRouter(
+  agents: Map<UUID, IAgentRuntime>,
+  serverInstance?: AgentServer
+): express.Router {
+  const router = express.Router();
+
+  // Get Agent Panels (public GET routes)
+  router.get('/:agentId/panels', async (req, res) => {
+    const agentId = validateUuid(req.params.agentId);
+    if (!agentId) {
+      return sendError(res, 400, 'INVALID_ID', 'Invalid agent ID format');
+    }
+
+    const runtime = agents.get(agentId);
+    if (!runtime) {
+      return sendError(res, 404, 'NOT_FOUND', 'Agent not found');
+    }
+
+    try {
+      const publicPanels = runtime.plugins
+        .flatMap((plugin) => plugin.routes || [])
+        .filter((route) => route.public === true && route.type === 'GET' && route.name)
+        .map((route) => ({
+          name: route.name,
+          // Construct the full path to the plugin route
+          path: `/api/agents/${agentId}/plugins/${runtime.plugins.find((p) => p.routes?.includes(route))?.name}${route.path.startsWith('/') ? route.path : `/${route.path}`}`,
+        }));
+
+      sendSuccess(res, publicPanels);
+    } catch (error) {
+      logger.error(`[AGENT PANELS] Error retrieving panels for agent ${agentId}:`, error);
+      sendError(res, 500, 'PANEL_ERROR', 'Error retrieving agent panels', error.message);
+    }
+  });
+
+  return router;
+}
