+++ packages/cli/src/server/api/shared/index.ts
@@ -1,0 +1,6 @@
+// Shared utilities index - provides a single entry point for all shared utilities
+export * from './response-utils';
+export * from './validation';
+export * from './file-utils';
+export * from './middleware';
+export * from './uploads';
