+++ packages/cli/src/server/api/shared/uploads/index.ts
@@ -1,0 +1,3 @@
+// Re-export upload configurations from the main upload module
+// This allows us to import upload configs from the shared directory
+export { agentUpload, channelUpload } from '../../../upload';
