--- packages/cli/src/server/test/api-routes.test.ts
@@ -1,438 +1,0 @@
-#!/usr/bin/env node
-
-import { AgentServer } from '../index';
-import { v4 as uuidv4 } from 'uuid';
-import { type UUID, type Plugin, ChannelType } from '@elizaos/core';
-import type { Character } from '@elizaos/core';
-import { startAgent } from '../../commands/start';
-import sqlPlugin from '@elizaos/plugin-sql';
-import fs from 'fs';
-import path from 'path';
-
-// Test character for agent
-const testCharacter: Character = {
-  id: uuidv4() as UUID,
-  name: 'TestAgent',
-  bio: ['A test agent for API route testing'],
-  settings: {
-    model: 'gpt-3.5-turbo',
-    secrets: {},
-  },
-  plugins: [sqlPlugin.name],
-};
-
-// Simple HTTP test utility
-class TestClient {
-  private baseUrl: string;
-
-  constructor(private port: number) {
-    this.baseUrl = `http://localhost:${port}`;
-  }
-
-  async request(method: string, path: string, data?: any) {
-    const url = `${this.baseUrl}${path}`;
-    const options: RequestInit = {
-      method,
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    };
-
-    if (data && method !== 'GET') {
-      options.body = JSON.stringify(data);
-    }
-
-    const response = await fetch(url, options);
-    const body = await response.json().catch(() => null);
-
-    return {
-      status: response.status,
-      body,
-    };
-  }
-
-  post(path: string, data?: any) {
-    return this.request('POST', path, data);
-  }
-
-  get(path: string, query?: Record<string, any>) {
-    const queryString = query ? '?' + new URLSearchParams(query).toString() : '';
-    return this.request('GET', path + queryString);
-  }
-
-  delete(path: string) {
-    return this.request('DELETE', path);
-  }
-}
-
-async function runTests() {
-  console.log('🚀 Starting API Route Tests...\n');
-
-  let server: AgentServer;
-  let testClient: TestClient;
-  let testAgentId: UUID;
-  let testServerId: UUID;
-  let testChannelId: UUID;
-  let testUserId: UUID = uuidv4() as UUID;
-
-  try {
-    // Initialize server
-    console.log('📦 Initializing server...');
-    server = new AgentServer();
-
-    // Assign the startAgent function to the server
-    server.startAgent = async (character) => {
-      return startAgent(character, server, undefined, [sqlPlugin]);
-    };
-
-    await server.initialize({
-      dataDir: './test-data',
-    });
-
-    // Start server on test port
-    const testPort = 3457;
-    await new Promise<void>((resolve, reject) => {
-      try {
-        server.start(testPort);
-        server.server.once('listening', () => {
-          console.log(`✅ Server started on port ${testPort}\n`);
-          resolve();
-        });
-        server.server.once('error', (err: any) => {
-          if (err.code === 'EADDRINUSE') {
-            console.error(
-              `❌ Port ${testPort} is already in use. Please stop any running servers.`
-            );
-          }
-          reject(err);
-        });
-      } catch (err) {
-        reject(err);
-      }
-    });
-
-    // Create test client
-    testClient = new TestClient(testPort);
-
-    // Create and start test agent
-    console.log('🤖 Creating test agent...');
-    const agent = await server.database.ensureAgentExists(testCharacter);
-    testAgentId = agent.id;
-    await server.startAgent(agent);
-    console.log(`✅ Test agent created with ID: ${testAgentId}\n`);
-
-    // Test agent list endpoint
-    console.log('📝 Testing agent list endpoint...');
-    const agentsListResponse = await testClient.get('/api/agents');
-    console.log(`  GET /api/agents - Status: ${agentsListResponse.status}`);
-    if (agentsListResponse.status !== 200) {
-      throw new Error(`Failed to get agents list: ${JSON.stringify(agentsListResponse.body)}`);
-    }
-    console.log(`  Found ${agentsListResponse.body.data.agents.length} agent(s)`);
-    console.log('✅ Agent list endpoint working\n');
-
-    // Test 1: Basic API endpoints
-    console.log('📝 Test 1: Testing basic API endpoints...');
-
-    const helloResponse = await testClient.get('/api/hello');
-    console.log(`  GET /api/hello - Status: ${helloResponse.status}`);
-    if (helloResponse.status !== 200) {
-      throw new Error(`Hello endpoint failed: ${JSON.stringify(helloResponse.body)}`);
-    }
-
-    const statusResponse = await testClient.get('/api/status');
-    console.log(`  GET /api/status - Status: ${statusResponse.status}`);
-    if (statusResponse.status !== 200) {
-      throw new Error(`Status endpoint failed: ${JSON.stringify(statusResponse.body)}`);
-    }
-
-    const pingResponse = await testClient.get('/api/ping');
-    console.log(`  GET /api/ping - Status: ${pingResponse.status}`);
-    if (pingResponse.status !== 200) {
-      throw new Error(`Ping endpoint failed: ${JSON.stringify(pingResponse.body)}`);
-    }
-    console.log('✅ Basic API endpoints working\n');
-
-    // Test 2: Messages API - Get Servers (should be empty initially)
-    console.log('📝 Test 2: Testing Messages API - Servers...');
-
-    const getServersResponse = await testClient.get('/api/messages/central-servers');
-    console.log(`  GET /api/messages/central-servers - Status: ${getServersResponse.status}`);
-    if (getServersResponse.status !== 200) {
-      throw new Error(`Get servers failed: ${JSON.stringify(getServersResponse.body)}`);
-    }
-    console.log(`  Found ${getServersResponse.body.data.servers.length} servers`);
-    console.log('✅ Central servers endpoint working\n');
-
-    // Test 3: Create Server
-    console.log('📝 Test 3: Creating server...');
-    const createServerResponse = await testClient.post('/api/messages/servers', {
-      name: 'Test Server',
-      sourceType: 'test',
-      sourceId: 'test-server-1',
-      metadata: { test: true },
-    });
-
-    console.log(`  POST /api/messages/servers - Status: ${createServerResponse.status}`);
-    if (createServerResponse.status !== 201) {
-      throw new Error(`Failed to create server: ${JSON.stringify(createServerResponse.body)}`);
-    }
-    testServerId = createServerResponse.body.data.server.id;
-    console.log(`✅ Central server created with ID: ${testServerId}\n`);
-
-    // Associate test agent with test server
-    console.log('🔗 Associating agent with server...');
-    await server.addAgentToServer(testServerId, testAgentId);
-    console.log(`✅ Agent associated with server\n`);
-
-    // Test 4: Create Central Channel
-    console.log('📝 Test 4: Creating central channel...');
-    const createChannelResponse = await testClient.post('/api/messages/channels', {
-      messageServerId: testServerId,
-      name: 'Test Channel',
-      type: ChannelType.GROUP,
-      metadata: { test: true },
-    });
-
-    console.log(`  POST /api/messages/channels - Status: ${createChannelResponse.status}`);
-    if (createChannelResponse.status !== 201) {
-      throw new Error(`Failed to create channel: ${JSON.stringify(createChannelResponse.body)}`);
-    }
-    testChannelId = createChannelResponse.body.data.channel.id;
-    console.log(`✅ Central channel created with ID: ${testChannelId}\n`);
-
-    // Test 5: Submit Message
-    console.log('📝 Test 5: Submitting message to central store...');
-    const submitMessageResponse = await testClient.post('/api/messages/submit', {
-      channel_id: testChannelId,
-      server_id: testServerId,
-      author_id: testUserId,
-      content: 'Hello from test user',
-      source_type: 'test',
-      metadata: {
-        agentName: 'Test User',
-      },
-    });
-
-    console.log(`  POST /api/messages/submit - Status: ${submitMessageResponse.status}`);
-    if (submitMessageResponse.status !== 201) {
-      throw new Error(`Failed to submit message: ${JSON.stringify(submitMessageResponse.body)}`);
-    }
-    const messageId = submitMessageResponse.body.data.id;
-    console.log(`✅ Message submitted with ID: ${messageId}\n`);
-
-    // Test 6: Get Messages for Channel
-    console.log('📝 Test 6: Retrieving messages from channel...');
-    const getMessagesResponse = await testClient.get(
-      `/api/messages/central-channels/${testChannelId}/messages`
-    );
-
-    console.log(
-      `  GET /api/messages/central-channels/${testChannelId}/messages - Status: ${getMessagesResponse.status}`
-    );
-    if (getMessagesResponse.status !== 200) {
-      throw new Error(`Failed to retrieve messages: ${JSON.stringify(getMessagesResponse.body)}`);
-    }
-    console.log(`✅ Retrieved ${getMessagesResponse.body.data.messages.length} messages\n`);
-
-    // Test 7: Get Channel Details
-    console.log('📝 Test 7: Getting channel details...');
-    const getChannelDetailsResponse = await testClient.get(
-      `/api/messages/central-channels/${testChannelId}/details`
-    );
-
-    console.log(
-      `  GET /api/messages/central-channels/${testChannelId}/details - Status: ${getChannelDetailsResponse.status}`
-    );
-    if (getChannelDetailsResponse.status !== 200) {
-      throw new Error(
-        `Failed to get channel details: ${JSON.stringify(getChannelDetailsResponse.body)}`
-      );
-    }
-    console.log(`✅ Retrieved channel details for: ${getChannelDetailsResponse.body.data.name}\n`);
-
-    // Test 8: Get Channel Participants
-    console.log('📝 Test 8: Getting channel participants...');
-    const getParticipantsResponse = await testClient.get(
-      `/api/messages/central-channels/${testChannelId}/participants`
-    );
-
-    console.log(
-      `  GET /api/messages/central-channels/${testChannelId}/participants - Status: ${getParticipantsResponse.status}`
-    );
-    if (getParticipantsResponse.status !== 200) {
-      throw new Error(
-        `Failed to get participants: ${JSON.stringify(getParticipantsResponse.body)}`
-      );
-    }
-    console.log(`✅ Retrieved ${getParticipantsResponse.body.data.length} participants\n`);
-
-    // Test 9: Delete Message
-    console.log('📝 Test 9: Deleting message...');
-    const deleteMessageResponse = await testClient.delete(
-      `/api/messages/central-channels/${testChannelId}/messages/${messageId}`
-    );
-
-    console.log(
-      `  DELETE /api/messages/central-channels/${testChannelId}/messages/${messageId} - Status: ${deleteMessageResponse.status}`
-    );
-    if (deleteMessageResponse.status !== 204) {
-      throw new Error(`Failed to delete message: ${JSON.stringify(deleteMessageResponse.body)}`);
-    }
-    console.log(`✅ Message deleted successfully\n`);
-
-    // Test 10: Agent Direct Message API
-    console.log('📝 Test 10: Testing agent direct message API...');
-    const agentMessageResponse = await testClient.post(`/api/agents/${testAgentId}/message`, {
-      channelId: testChannelId,
-      serverId: testServerId,
-      entityId: testUserId,
-      text: 'Direct message to agent',
-      userName: 'Test User',
-      source: 'test',
-    });
-
-    console.log(
-      `  POST /api/agents/${testAgentId}/message - Status: ${agentMessageResponse.status}`
-    );
-    if (agentMessageResponse.status !== 202) {
-      throw new Error(`Failed to send agent message: ${JSON.stringify(agentMessageResponse.body)}`);
-    }
-    console.log(`✅ Agent message accepted for processing\n`);
-
-    // Test 11: Create DM Channel
-    console.log('📝 Test 11: Creating DM channel...');
-    const user1Id = uuidv4() as UUID;
-    const user2Id = uuidv4() as UUID;
-
-    const dmChannelResponse = await testClient.get('/api/messages/dm-channel', {
-      currentUserId: user1Id,
-      targetUserId: user2Id,
-      dmServerId: testServerId,
-    });
-
-    console.log(`  GET /api/messages/dm-channel - Status: ${dmChannelResponse.status}`);
-    if (dmChannelResponse.status !== 200) {
-      throw new Error(`Failed to create DM channel: ${JSON.stringify(dmChannelResponse.body)}`);
-    }
-    console.log(`✅ DM channel created: ${dmChannelResponse.body.data.name}\n`);
-
-    // Test 12: Create Group Channel with Participants
-    console.log('📝 Test 12: Creating group channel with participants...');
-    const groupChannelResponse = await testClient.post('/api/messages/central-channels', {
-      name: 'Test Group Channel',
-      participantCentralUserIds: [user1Id, user2Id, testUserId],
-      type: ChannelType.GROUP,
-      server_id: testServerId,
-      metadata: { test: true },
-    });
-
-    console.log(`  POST /api/messages/central-channels - Status: ${groupChannelResponse.status}`);
-    if (groupChannelResponse.status !== 201) {
-      throw new Error(
-        `Failed to create group channel: ${JSON.stringify(groupChannelResponse.body)}`
-      );
-    }
-    console.log(`✅ Group channel created: ${groupChannelResponse.body.data.name}\n`);
-
-    // Test 13: Server-Agent Association APIs
-    console.log('📝 Test 13: Testing server-agent association APIs...');
-
-    // Get agents for server
-    const getAgentsResponse = await testClient.get(`/api/messages/servers/${testServerId}/agents`);
-    console.log(
-      `  GET /api/messages/servers/${testServerId}/agents - Status: ${getAgentsResponse.status}`
-    );
-    if (getAgentsResponse.status !== 200) {
-      throw new Error(`Failed to get agents for server: ${JSON.stringify(getAgentsResponse.body)}`);
-    }
-    console.log(`  Found ${getAgentsResponse.body.data.agents.length} agents in server`);
-
-    // Get servers for agent
-    const getAgentServersResponse = await testClient.get(
-      `/api/messages/agents/${testAgentId}/servers`
-    );
-    console.log(
-      `  GET /api/messages/agents/${testAgentId}/servers - Status: ${getAgentServersResponse.status}`
-    );
-    if (getAgentServersResponse.status !== 200) {
-      throw new Error(
-        `Failed to get servers for agent: ${JSON.stringify(getAgentServersResponse.body)}`
-      );
-    }
-    console.log(`  Agent belongs to ${getAgentServersResponse.body.data.servers.length} servers`);
-
-    // Test remove and re-add agent to server
-    const removeAgentResponse = await testClient.delete(
-      `/api/messages/servers/${testServerId}/agents/${testAgentId}`
-    );
-    console.log(
-      `  DELETE /api/messages/servers/${testServerId}/agents/${testAgentId} - Status: ${removeAgentResponse.status}`
-    );
-    if (removeAgentResponse.status !== 200) {
-      throw new Error(
-        `Failed to remove agent from server: ${JSON.stringify(removeAgentResponse.body)}`
-      );
-    }
-
-    const addAgentResponse = await testClient.post(`/api/messages/servers/${testServerId}/agents`, {
-      agentId: testAgentId,
-    });
-    console.log(
-      `  POST /api/messages/servers/${testServerId}/agents - Status: ${addAgentResponse.status}`
-    );
-    if (addAgentResponse.status !== 201) {
-      throw new Error(`Failed to add agent to server: ${JSON.stringify(addAgentResponse.body)}`);
-    }
-    console.log(`✅ Server-agent association APIs working\n`);
-
-    // Test 14: Channel Media Upload
-    console.log('📝 Test 14: Testing channel media upload...');
-
-    // Create a test file
-    const testImagePath = path.join(process.cwd(), 'test-image.png');
-    const testImageData = Buffer.from(
-      'iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNkYPhfDwAChwGA60e6kgAAAABJRU5ErkJggg==',
-      'base64'
-    );
-    fs.writeFileSync(testImagePath, testImageData);
-
-    // Note: File upload requires multipart/form-data which our simple test client doesn't support
-    // We'll just verify the endpoint exists by checking with a GET request
-    console.log(`  Skipping actual file upload test (requires multipart/form-data)`);
-
-    // Clean up test file
-    fs.unlinkSync(testImagePath);
-    console.log(`✅ Channel media upload endpoint exists\n`);
-
-    // Summary
-    console.log('📊 Test Summary:');
-    console.log('  ✅ Basic API endpoints - PASSED');
-    console.log('  ✅ Central servers endpoint - PASSED');
-    console.log('  ✅ Create server - PASSED');
-    console.log('  ✅ Create channel - PASSED');
-    console.log('  ✅ Submit message - PASSED');
-    console.log('  ✅ Get messages - PASSED');
-    console.log('  ✅ Get channel details - PASSED');
-    console.log('  ✅ Get participants - PASSED');
-    console.log('  ✅ Delete message - PASSED');
-    console.log('  ✅ Agent direct message - PASSED');
-    console.log('  ✅ Create DM channel - PASSED');
-    console.log('  ✅ Create group channel - PASSED');
-    console.log('  ✅ Server-agent associations - PASSED');
-    console.log('\n✅ All tests passed successfully! 🎉\n');
-  } catch (error) {
-    console.error('❌ Test failed:', error);
-    process.exit(1);
-  } finally {
-    // Cleanup
-    if (server) {
-      console.log('🧹 Cleaning up...');
-      await server.stop();
-      console.log('✅ Server stopped\n');
-    }
-  }
-}
-
-// Run the tests
-runTests().catch(console.error);
