--- packages/cli/src/server/test/run-api-tests.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-# Run API Route Tests
-echo "🚀 Running API Route Tests..."
-cd "$(dirname "$0")/../../.."
-npx tsx src/server/test/api-routes.test.ts
