--- packages/cli/src/server/test/verify-frontend-fix.ts
@@ -1,140 +1,0 @@
-#!/usr/bin/env node
-
-console.log('🧪 Verifying Frontend Loading Fix');
-console.log('==================================\n');
-
-const serverUrl = 'http://localhost:3000';
-const agentId = 'b850bc30-45f8-0041-a00a-83df46d8555d';
-
-async function verifyServerIsRunning() {
-  console.log('📡 Checking if server is running...');
-  try {
-    const response = await fetch(`${serverUrl}/api/ping`);
-    const data = await response.json();
-    if (data.pong) {
-      console.log('  ✅ Server is running');
-      return true;
-    } else {
-      console.log('  ❌ Server ping failed');
-      return false;
-    }
-  } catch (error) {
-    console.log('  ❌ Server not responding');
-    console.log('  💡 Please start the server with: npm start');
-    return false;
-  }
-}
-
-async function verifyAgent() {
-  console.log('\n👤 Checking agent...');
-  try {
-    const response = await fetch(`${serverUrl}/api/agents/${agentId}`);
-    const data = await response.json();
-    if (data.success) {
-      console.log(`  ✅ Agent found: ${data.data.name} (${data.data.status})`);
-      return true;
-    } else {
-      console.log('  ❌ Agent not found');
-      return false;
-    }
-  } catch (error) {
-    console.log('  ❌ Error checking agent:', error.message);
-    return false;
-  }
-}
-
-async function verifyAPIEndpoints() {
-  console.log('\n🔌 Checking API endpoints...');
-
-  // Check servers endpoint
-  try {
-    const response = await fetch(`${serverUrl}/api/messages/central-servers`);
-    const data = await response.json();
-    console.log(`  ✅ Central servers API: ${data.success ? 'Working' : 'Error'}`);
-  } catch (error) {
-    console.log('  ❌ Central servers API failed');
-  }
-
-  // Ensure we have at least one server
-  try {
-    const response = await fetch(`${serverUrl}/api/messages/servers`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        name: 'Test Server for Frontend Fix',
-        sourceType: 'test_fix',
-      }),
-    });
-    const data = await response.json();
-    if (data.success) {
-      console.log('  ✅ Test server created for verification');
-      return data.data.server.id;
-    }
-  } catch (error) {
-    console.log('  ⚠️  Could not create test server, but this may not be needed');
-  }
-
-  return null;
-}
-
-function explainFix() {
-  console.log('\n🔧 What Was Fixed:');
-  console.log('==================\n');
-
-  console.log('✅ BEFORE (Broken):');
-  console.log('  • URL: /chat/agentId');
-  console.log('  • Component expected: channelId (path) + agentId & serverId (query)');
-  console.log('  • Result: "Loading chat context..." forever');
-  console.log('');
-
-  console.log('✅ AFTER (Fixed):');
-  console.log('  • URL: /chat/agentId (same URL)');
-  console.log('  • Component now detects "agent mode"');
-  console.log('  • Automatically creates DM channel & server');
-  console.log('  • Chat loads successfully!');
-  console.log('');
-
-  console.log('🎯 How It Works Now:');
-  console.log('  1. User visits /chat/agentId');
-  console.log('  2. Component detects agent mode (agentId in path, no query params)');
-  console.log('  3. Component auto-creates/finds server & DM channel');
-  console.log('  4. Chat context loads with all required IDs');
-  console.log('  5. User can type and chat works!');
-}
-
-function showTestInstructions() {
-  console.log('\n🧪 How to Test:');
-  console.log('================\n');
-
-  console.log('1. Open your browser');
-  console.log('2. Navigate to:');
-  console.log(`   ${serverUrl}/chat/${agentId}`);
-  console.log('3. You should see:');
-  console.log('   ✅ Chat interface loads (no "Loading chat context...")');
-  console.log('   ✅ Input field is enabled');
-  console.log('   ✅ You can type messages');
-  console.log('   ✅ Agent responds to your messages');
-  console.log('');
-  console.log('4. If you still see "Loading chat context...":');
-  console.log('   • Clear browser cache (Ctrl+Shift+R)');
-  console.log('   • Try incognito/private window');
-  console.log('   • Check browser console for errors');
-}
-
-async function runVerification() {
-  const serverOk = await verifyServerIsRunning();
-  if (!serverOk) return;
-
-  const agentOk = await verifyAgent();
-  if (!agentOk) return;
-
-  await verifyAPIEndpoints();
-
-  explainFix();
-  showTestInstructions();
-
-  console.log('\n🎉 Frontend loading fix deployed successfully!');
-  console.log('   The chat should now work at: /chat/agentId URLs');
-}
-
-runVerification().catch(console.error);
